// react-native-ios-android-web/packages/app/data/coinGeckoClient/types.ts
export type Market = {
    id: string,
    symbol: string
    name: string
    image?: string
    current_price?: number
    market_cap?: number
    market_cap_rank?: number
    fully_diluted_valuation?: number
    total_volume?: number
    high_24h?: number
    low_24h?: number
    price_change_24h?: number
    price_change_percentage_24h?: number
    market_cap_change_24h?: number
    market_cap_change_percentage_24h?: number
    circulating_supply?: number
    total_supply?: number
    max_supply?: number
    ath?: number
    ath_change_percentage?: number
    ath_date?: string
    atl?: number
    atl_change_percentage?: number
    atl_date?: string
    last_updated?: string,
    sparkline_in_7d: {
        price: Array<number>
    }
}

export type Coin = {
    id: string
    name: string
    api_symbol: string
    symbol: string
    market_cap_rank?: number
    thumb?: string
    large?: string
}

export type NFT = {
    id: string
    name: string
    symbol: string
    thumb?: string
}

export type SearchResult = {
    coins: Array<Coin>
    nfts: Array<NFT>
}